Fixed Compilation Error
Close #1
Since C++11 it is not necessary to add spaces between closing brackets
of template arguments. In the spirit of compatibility however, the
spaces between those brackets have been added.
In general, more compatibility issues were removed.
A patch file and a shell script helping with the installation of gigist
in an existing or new cpptraj installation is supplied. Usage for the
shell script:
bash patch.sh </ROUTE/TO/CPPTRAJ/DIRECTORY>
If the route to the cpptraj diorectory is not supplied, uses the cpptraj
installation from AMBERHOME.