Add functionality for COM
Since the Center of Mass is the most usefull metric to decide, where a
certain atom should be gridded at (I guess), implementation was started
with this commit.
Plan is to calculate the COM once with reference to the first atom in a
molecule and then turn it with the molecule (Quaternion rotation that is
already implemented). And then calculate the position on the grid.