Added COM calculation
Binning can now be performed using the COM of the solvent molecule. This
is helpfull when using solvent molecules that are rigid but do not have
a distinct center atom, e.g., Benzene.
For this code was embedded in the Action_GIGist class, to calculate the
COM. Also, the calculation of the Quaternions was moved to a different
function and is now calculated there.
In the Quaternion class, an empty constructor is added, to be able to
build the Quaternion before it is assigned a true value.
All new functions are declared in the file Action_GIGIST.h.