Fixes for C++11
Started to make the code conform to the C++11 standard, this is done to
adhere to usual standards and have a solid code base here.
Changed setting the outfilename to the function GetStringKey, in order
to have a more consistent access to the different parameters from the
command line prompt.