Fixed formatting issues
Fixed problems with tabs and spaces.
In function descriptions, fixed the name argument to param, as is
standard for function descriptions.
Added a check if HUGE is defined, if not defines HUGE in accordance with
SVID (compare with math.h).
In all CUDA Kernel functions, changed the Test class to Coordinates_GPU,
in order to make clear what this class actually represents.
Fixed include statements to always include the header files for C++. For
C-Headers, use the c before the name, e.g., #include <cmath> instead of
the C-like #inlcude <math.h>. This also is more compliant with C++
standards.