Fixed error with Ions.
I never considered the ions being in the topology file before the water
molecules. Therefor, I did not check if the water atoms are actually in
front in the topology file, so now the implementation checks whether the
molecule in question actually has at least 3 atoms (so that a quaternion
is possible).