Grouped the information for the GIST calculation
I created a struct (containing different other structs), in the
Action_GIGist class. This allows for a more elegant access to the
information needed in the GIST calculation. Of course, for this, some
refactoring had to be done.
But the new format allows for an easier understanding, as well as for an
easier access.
I also incorporated the FEBISS code from Miguel Steiner, which allows
for a sophisticated water placement. Also some of this code had to be
changed due to the refactoring of the data structure.
I also changed the solvent_ array to a smart pointer (std::unique_ptr),
to allow for safer usage and access.
Finally, a bit of code in the CUDA kernel had to be changed, as this was
not working properly.
Also, in the setup a bit of refactoring was done, pulling out things
that belong together into functions, this will continue. The
documentation of this code is not yet complete, but will be done, as
soon as the refactoring of the init and setup functions is finished.