Lots of refactoring
Refactored many parts of the GIST setup and initialization, also started
with refactoring of the DoAction function. This function is still MUCH
too long and will be made smaller.
Also changed a few things, I removed the min and max pointers, as these
were not used anyway.
The refactored functions must still be documented in the header file, as
well as will be moved to the end of the cpp file. This allows a better
understanding of the code and what happens at which point in the
calculation.